[build][ulib][kernel] Allow header-only use of standard C++ library
The libc++ implementation of the standard C++ library is compatible
with both our Clang and GCC compilers.  We get the headers from the
Clang compiler prebuilt even when building with GCC.

Permit limited header-only use in Zircon code.  To start with, only
the <new> header is used, replacing <zxcpp/new.h> and <fbl/new.h>.

Bug: ZX-1751 #comment Build system support for header-only; <new> header used
Test: CQ
Change-Id: Iee8cea0b637e4eb9edc6207af1e5098a6365effd