[syscalls] Refactor hyp syscall interface
Instead of using a single Zircon syscall for hypervisor call
forwarding, we now use several syscalls which operate on generic
hypervisor constructs (currently virtual IRQs and pipes/message
queues). This provides an interface for user-level applications and
drivers that can be independent of the underlying hypervisor.

This patch also updates the link-shbuf and vs-shbuf drivers to use
this new interface.