[devmgr] Cancel waiters when unbinding children
If a device driver performs a device_state_wait() on its parent, and
the parent starts to unbind its children, the wait should be canceled
immediately, otherwise the wait can continue after the child's release
hook is called. This has been changed so that an internal signal is
asserted when children are unbound, waking any that are waiting.